[REFACTOR] extension 디렉토리 생성 Color extension 이동
