+++ POMPOM/POMPOM/UI/ClothPickerView/ViewModel/PickerViewModel.swift
@@ -60,25 +60,6 @@ class PickerViewModel: ObservableObject {
     }
 }
 
-extension Color {
-    init(hex: String) {
-        let scanner = Scanner(string: hex)
-        scanner.currentIndex = hex.startIndex
-        var rgbValue: UInt64 = 0
-        scanner.scanHexInt64(&rgbValue)
-
-        let r = (rgbValue & 0xff0000) >> 16
-        let g = (rgbValue & 0xff00) >> 8
-        let b = rgbValue & 0xff
-
-        self.init(
-            red: CGFloat(r) / 0xff,
-            green: CGFloat(g) / 0xff,
-            blue: CGFloat(b) / 0xff
-        )
-    }
-}
-
 struct Cloth: Identifiable {
     var id: Int
     var name: String
