+++ POMPOM/POMPOM/Util/Extensions/Color.swift
@@ -1,0 +1,26 @@
+//
+//  Color.swift
+//  POMPOM
+//
+//  Created by GOngTAE on 2022/06/14.
+//
+
+import SwiftUI
+
+extension Color {
+    init(hex: String) {
+        let scanner = Scanner(string: hex)
+        scanner.currentIndex = hex.startIndex
+        var rgbValue: UInt64 = 0
+        scanner.scanHexInt64(&rgbValue)
+
+        let r = (rgbValue & 0xff0000) >> 16
+        let g = (rgbValue & 0xff00) >> 8
+        let b = rgbValue & 0xff
+
+        self.init(
+            red: CGFloat(r) / 0xff,
+            green: CGFloat(g) / 0xff,
+            blue: CGFloat(b) / 0xff
+        )
+    }
